extracted class
